refactor: drop redundant attach.ts
