+++ packages/neovim/src/index.ts
@@ -1,4 +1,4 @@
-export { attach } from './attach';
+export { attach } from './attach/attach';
 export { Neovim, NeovimClient, Buffer, Tabpage, Window } from './api/index';
 export { Plugin, Function, Autocmd, Command } from './plugin';
 export { NvimPlugin } from './host/NvimPlugin';
